requirements.txt
mediapipe==0.10.20
numpy==1.26.0
opencv_contrib_python==4.11.0.86
opencv_python_headless==4.10.0.84
Pillow==11.1.0
streamlit==1.40.2
